[nextest] rename output.rs to test_list/output_format.rs
Make way for output options to be added soon.
